requirements.txt
Django>=4.2
tensorflow
numpy
pandas
pillow
opencv-python
matplotlib
scikit-learn
google-generativeai
python-dotenv
